backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
clerk-sdk-python
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.1
